requirements.txt
Flask==3.1.3
python-dotenv>=1.0,<2.0
waitress>=3.0,<4.0
